Working Version
Basic feature and functionality implemented. Program is still very rough
and needs refinement.

This version was completed a few days after the last commit (almost a
month prior to this commit). I thought I had synced the version to
github, but I had not. I started to fiddle with the files, and ended up
accidentally deleting them. After a long time of recovering and fixing
the corrupted files, along with doing my other work, I finally got to
restore the version of the program just before the accident.

There seems to be a problem with the merging of the old files with the
recovered files. I'll be creating this new project instead of updating
the old Basic-Financial-Tracker project.